// phoenix-common-react(1)/src/types/LCL/booking/booking.types.ts
export interface Clause {
  clauseCode: string;
  clause?: string | null;
  clauseDesc?: string;
  clauseDescLocale?: string;
  clauseLocale?: string;
  clauseName?: string;
  clauseNameLocale?: string;
  sequence?: number;
}
// ...
export interface BookingQuoteBean {
  referenceNumber: number | null;
  bookQuoteDate: string | null;
  bookingQuoteType: string;
  handlingOffice: string;
  quoteNumber: number | null;
  userReference: string;
  takenBy: string;
  status: string;
  transmitToLocation1: string;
  terms: string;
  pickupNeeded: YesNo;
  lineBooking: string;
  lineCode: string;
  numberOfPieces: number;
  ratingType: string;
  discount: number;
  prepaidCredit: string;
  aesAmsFlag: YesNo;
  receivedVia: string;
  vesselVoyageID: number;
  shipmendID: number;
  frequency: string;
  transitTime: string;
  receivedFromName: string;
  tranShipmentFlag: YesNo;
  flashPoint: number;
  container1: number;
  hazardousAction: string;
  bookingQuoteCustomerBean: BookingQuoteCustomerBean;
  actualBuyerSellerBean: Record<string, unknown>;
  bookingQuoteRoutingBean: BookingQuoteRoutingBean;
  bookingQuoteCargoBean: BookingQuoteCargoBean;
  bookingEquipmentBean: Record<string, unknown>;
  shipmentStatusUpdateBean: ShipmentStatusUpdateBean;
  bookingQuoteCustomFilingBean: BookingQuoteCustomFilingBean;
  type: string;
  posTimeFormat: number;
  controlFlag: string;
  preliminaryBooking: boolean;
  preliminaryBookingStatus: string;
  termName: string;
  nomination: string;
  rateControllingEntity: string;
  updatedBy: string;
  isCTCChargeFromQuote: boolean;
  aesItnHoldStatusId: number;
  ucrHoldStatusId: number;
  isOldBooking: boolean;
  commission: string;
  carriers: unknown[];
  clauses: Clause[];
  bookingCustomerDeclaredCargoBeanList: BookingCustomerDeclaredCargoBean[];
  roeType: string;
  includePlcOnDocument: boolean;
  transmissionService: string;
  transmitToLocationName: string;
  purchaseOrderReferenceNumber: string;
  invoiceCurrency: string;
  invoiceToLocalExchangeRate: number;
  oldReferenceNumber: number;
  oldStatus: string;
  modifyPLC: boolean;
  customerChange: boolean;
  isUpdated: boolean;
  isShipmentOrderTransmit: boolean;
  isTransmitted: boolean;
  isCancelSOTransmitted: boolean;
  isVip: boolean;
  isAutoGeneratedDeliveryReferenceExist: boolean;
  thirdPartySystem: string;
  isManualRemoveCustomsHold: boolean;
  isLotCreatedForBKG: boolean;
  isBolCreatedForBKG: boolean;
  isCustomerOwnCFSAgreement: YesNo;
  isEdiBooking: boolean;
  modeOfTransport: string;
  directLoading: YesNo;
  tInputDate: string;
  errorMessage: string;
  isSensitiveCargoApprovalReceived: boolean;
  direction: string;
  vid: YesNo;
  importBookingNumber: number;
  bookingCarrierBean: Record<string, unknown>;
  isIntraCarrier: boolean;
  bookingEdiBeans: unknown[];
  isPrintDimension: boolean;
  fillingByMrn: string;
  bookingQuoteRoutingBeanlist: BookingQuoteRoutingBean[];
  masterQuoteNumber: number;
  isQuoteChild: boolean;
  onDock: YesNo;
  bookingQuoteMultiCargoBeanList: BookingQuoteMultiCargoBean[];
  multiCargoBookingQuoteNoteBeans: MultiCargoBookingQuoteNoteBean[];
  cargoHsCodeNoteBeans: unknown[];
  isRoutingUpdated: boolean;
  pendingFinalBookingStatus: YesNo;
  pendingFinalQuoteStatus: YesNo;
  hazaRuleNotes: string;
  pendingFinalBooking: boolean;
  isRatedNonRatedDocSend: boolean;
  isBookingTransmittedtoSaco: boolean;
  isBookingTruckingInstrSent: boolean;
  updateComments: string;
  deleteComments: string;
  updateExpenseComments: string;
  deleteExpenseComments: string;
  addedExpenseComments: string;
  multiCustomHoldStatus: string;
  isCarrCodeRestricted: boolean;
  originChatAppChannel: boolean;
  doorDeliveryDetailsBean: DoorDeliveryDetailsBean;
  cfsDeliveryType: string;
  customerTypeChange: boolean;
  isHazardousPermissionOverride: YesNo;
  pickupUpdatedComment: string;
  pickupDeleteComment: string;
  doorDeliveryDeletedComment: string;
  doorDeliveryUpdatedComment: string;
  amsCustomAdvanceFillingMainBean: AmsCustomAdvanceFillingMainBean;
  oldamsCustomAdvanceFillingMainBean: AmsCustomAdvanceFillingMainBean;
  isVesselUpdate: boolean;
  isMultipleLot: boolean;
  isLotCmtFromQuo: boolean;
  toggleForShipmentPickedUpEvent: boolean;
  routed: YesNo;
  isLegacy: boolean;
  isCOCUpdated: boolean;
  isBookingFileFrobCargo: boolean;
  tmsRateFetched: YesNo;
  customerOwnContainerToggle: YesNo;
  updatedOn: string | null;
  cfsLoadPort: string;
  paymentMethodCode: string;
  customManifestFeeCollectBy: string;
  specialHandlingCode: string;
  controlNumber: string;
  agentCode: string;
  siChannel: string;
  tmsShipmentId?: string;
  truckerProNumber?: string;
  transmit?:string;
  shipperEmail?:string;
  intertrackFlag?:string;
  documentCutoffTime?:string;
  wwaReference?:string;
}